+++ packages/amplify-codegen-e2e-tests/src/codegen-tests-base/datastore-modelgen.ts
@@ -1,6 +1,8 @@
 import {
     initProjectWithProfile,
-    addApiWithSchemaAndConflictDetection,
+    addApiWithBlankSchemaAndConflictDetection,
+    updateApiSchema,
+    createRandomName,
     generateModels,
     AmplifyFrontendConfig
 } from "amplify-codegen-e2e-core";
@@ -9,11 +11,14 @@ import path from 'path';
 import { isNotEmptyDir, generateSourceCode } from '../utils';
 
 export async function testCodegenModels(config: AmplifyFrontendConfig, projectRoot: string, schema: string) {
+    const name = createRandomName();
+
     // init project and add API category
-    await initProjectWithProfile(projectRoot, { ...config });
+    await initProjectWithProfile(projectRoot, { name, ...config });
 
     //enable datastore
-    await addApiWithSchemaAndConflictDetection(projectRoot, schema);
+    await addApiWithBlankSchemaAndConflictDetection(projectRoot);
+    await updateApiSchema(projectRoot, name, schema);
 
     //generate pre existing user file
     const userSourceCodePath = generateSourceCode(projectRoot, config.srcDir);