+++ packages/amplify-codegen-e2e-tests/src/codegen-tests-base/graphql-codegen.ts
@@ -1,9 +1,11 @@
 import {
     initProjectWithProfile,
-    addApiWithSchema,
+    addApiWithoutSchema,
+    updateApiSchema,
     addCodegen,
     AmplifyFrontendConfig,
-    generateStatementsAndTypes
+    generateStatementsAndTypes,
+    createRandomName
 } from "amplify-codegen-e2e-core";
 import { existsSync } from "fs";
 import path from 'path';
@@ -12,7 +14,9 @@ import { isNotEmptyDir, generateSourceCode } from '../utils';
 export async function testGraphQLCodegen(config: AmplifyFrontendConfig, projectRoot: string, schema: string) {
     // init project and add API category
     await initProjectWithProfile(projectRoot, { ...config });
-    await addApiWithSchema(projectRoot, schema);
+    const projectName = createRandomName();
+    await addApiWithoutSchema(projectRoot, { apiName: projectName });
+    await updateApiSchema(projectRoot, projectName, schema);
 
     // generate pre-existing user file
     const userSourceCodePath = generateSourceCode(projectRoot, config.srcDir);