+++ packages/appsync-modelgen-plugin/src/utils/process-connections-v2.ts
@@ -1,22 +1,24 @@
 import { CodeGenField, CodeGenFieldDirective, CodeGenModel, CodeGenModelMap } from '../visitors/appsync-visitor';
-import {
-  CodeGenFieldConnection, DEFAULT_HASH_KEY_FIELD, flattenFieldDirectives,
-  getDirective,
-  makeConnectionAttributeName,
-} from './process-connections';
+import { CodeGenFieldConnection, DEFAULT_HASH_KEY_FIELD, flattenFieldDirectives, makeConnectionAttributeName } from './process-connections';
 import { processHasOneConnection } from './process-has-one';
 import { processBelongsToConnection, getBelongsToConnectedField } from './process-belongs-to';
 import { processHasManyConnection } from './process-has-many';
+import { getDirective } from './fieldUtils';
 
 // TODO: This file holds several references to utility functions in the v1 process connections file, those functions need to go here before that file is removed
 
-export function getConnectedFieldV2(field: CodeGenField, model: CodeGenModel, connectedModel: CodeGenModel, directiveName: string): CodeGenField {
+export function getConnectedFieldV2(
+  field: CodeGenField,
+  model: CodeGenModel,
+  connectedModel: CodeGenModel,
+  directiveName: string,
+): CodeGenField {
   const connectionInfo = getDirective(field)(directiveName);
   if (!connectionInfo) {
     throw new Error(`The ${field.name} on model ${model.name} is not connected`);
   }
 
-  if(connectionInfo.name === 'belongsTo') {
+  if (connectionInfo.name === 'belongsTo') {
     let connectedFieldBelongsTo = getBelongsToConnectedField(field, model, connectedModel);
     if (connectedFieldBelongsTo) {
       return connectedFieldBelongsTo;
@@ -25,7 +27,7 @@ export function getConnectedFieldV2(field: CodeGenField, model: CodeGenModel, co
 
   const indexName = connectionInfo.arguments.indexName;
   const connectionFields = connectionInfo.arguments.fields;
-  if (connectionFields) {
+  if (connectionFields || directiveName === 'hasOne') {
     let indexDirective;
     if (indexName) {
       indexDirective = flattenFieldDirectives(connectedModel).find(dir => {
@@ -43,14 +45,23 @@ export function getConnectedFieldV2(field: CodeGenField, model: CodeGenModel, co
     }
 
     // when there is a fields argument in the connection
-    const connectedFieldName = indexDirective ? ((fieldDir: CodeGenFieldDirective) => { return fieldDir.fieldName ;})(indexDirective as CodeGenFieldDirective) : DEFAULT_HASH_KEY_FIELD;
+    const connectedFieldName = indexDirective
+      ? ((fieldDir: CodeGenFieldDirective) => {
+          return fieldDir.fieldName;
+        })(indexDirective as CodeGenFieldDirective)
+      : DEFAULT_HASH_KEY_FIELD;
 
     // Find a field on the other side which connected by a @connection and has the same fields[0] as indexName field
-    const otherSideConnectedField = connectedModel.fields.find(f => {
-      return f.directives.find(d => {
-        return (d.name === 'belongsTo' || d.name === 'hasOne' || d.name === 'hasMany') && d.arguments.fields && d.arguments.fields[0] === connectedFieldName;
-      });
-    });
+    const otherSideConnectedField = connectedModel.fields
+      .filter(f => f.type === model.name)
+      .find(f =>
+        f.directives.find(
+          d =>
+            (d.name === 'belongsTo' || d.name === 'hasOne' || d.name === 'hasMany') &&
+            d.arguments.fields &&
+            d.arguments.fields[0] === connectedFieldName,
+        ),
+      );
     if (otherSideConnectedField) {
       return otherSideConnectedField;
     }
@@ -58,7 +69,7 @@ export function getConnectedFieldV2(field: CodeGenField, model: CodeGenModel, co
     const connectedField = connectedModel.fields.find(f => f.name === connectedFieldName);
 
     if (!connectedField) {
-      throw new Error(`Can not find key field ${connectedFieldName} in ${connectedModel}`);
+      throw new Error(`Can not find key field ${connectedFieldName} in ${connectedModel.name}`);
     }
     return connectedField;
   }
@@ -68,14 +79,13 @@ export function getConnectedFieldV2(field: CodeGenField, model: CodeGenModel, co
   return connectedField
     ? connectedField
     : {
-      name: connectedFieldName,
-      directives: [],
-      type: 'ID',
-      isList: false,
-      isNullable: true,
-    };
+        name: connectedFieldName,
+        directives: [],
+        type: 'ID',
+        isList: false,
+        isNullable: true,
+      };
 }
-
 
 export function processConnectionsV2(
   field: CodeGenField,
@@ -84,9 +94,8 @@ export function processConnectionsV2(
 ): CodeGenFieldConnection | undefined {
   const connectionDirective = field.directives.find(d => d.name === 'hasOne' || d.name === 'hasMany' || d.name === 'belongsTo');
 
-  if(connectionDirective) {
-
-    switch(connectionDirective.name) {
+  if (connectionDirective) {
+    switch (connectionDirective.name) {
       case 'hasOne':
         return processHasOneConnection(field, model, modelMap, connectionDirective);
       case 'belongsTo':